Compiler warnings Error strings in vos_utils added
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk/trdp@1641 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ test/diverse/crc-test.c
@@ -8,12 +8,13 @@
  *
  * @author          Bernd Loehr, NewTec GmbH
  *
- * @remarks This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. 
+ * @remarks This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
  *          If a copy of the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  *          Copyright Bombardier Transportation Inc. or its subsidiaries and others, 2013. All rights reserved.
  *
  * $Id$
  *
+ *      BL 2017-06-30: Compiler warnings, local prototypes added
  */
 
 #include <stdio.h>
@@ -22,32 +23,36 @@
 
 #include "vos_utils.h"
 
-UINT8	gSampleDATA[] =
+UINT8 gSampleDATA[] =
 {
-	0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8,		/*	Testdata	*/
-    0x0, 0x0, 0x0, 0x0							/*	CRC	*/  
+    0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8,     /* Testdata	*/
+    0x0, 0x0, 0x0, 0x0                          /* CRC */
 };
 
-int main()
+int main ()
 {
-	// Compute CRC and store in little endian */
-    UINT32	myCrc; 
-    
-    myCrc = vos_crc32 (0, gSampleDATA, 8);
-    gSampleDATA[8] = myCrc;
-    gSampleDATA[9] = myCrc >> 8;
-    gSampleDATA[10] = myCrc >> 16;
-    gSampleDATA[11] = myCrc >> 24;
-    
+    /* Compute CRC and store in little endian * / */
+    UINT32 myCrc;
+
+    myCrc = vos_crc32(0, gSampleDATA, 8);
+    gSampleDATA[8]  = (UINT8) myCrc;
+    gSampleDATA[9]  = (UINT8) myCrc >> 8;
+    gSampleDATA[10] = (UINT8) myCrc >> 16;
+    gSampleDATA[11] = (UINT8) myCrc >> 24;
+
     printf("CRC: %08x\n", myCrc);
-    printf("LE in memory (from low to high): %02hhx %02hhx %02hhx %02hhx\n", gSampleDATA[8], gSampleDATA[9], gSampleDATA[10], gSampleDATA[11]);
-    
-	// CRC Check on Null
-    
-	// Verify CRC */
-    myCrc = vos_crc32 (0, gSampleDATA, 12);
+    printf("LE in memory (from low to high): %02hhx %02hhx %02hhx %02hhx\n",
+           gSampleDATA[8],
+           gSampleDATA[9],
+           gSampleDATA[10],
+           gSampleDATA[11]);
+
+    /* CRC Check on Null */
+
+    /* Verify CRC * / */
+    myCrc = vos_crc32(0, gSampleDATA, 12);
     printf("Checked CRC over all: %08x\n", myCrc);
-    
+
     if (myCrc == 0)
     {
         printf(" = Correct behavior!\n");
@@ -56,23 +61,27 @@ int main()
     {
         printf(" = Wrong!!\n");
     }
-    
-    // Compute CRC and store as big endian
-    myCrc = vos_crc32 (0, gSampleDATA, 8);
-    gSampleDATA[8] = myCrc >> 24;
-    gSampleDATA[9] = myCrc >> 16;
-    gSampleDATA[10] = myCrc >> 8;
-    gSampleDATA[11] = myCrc;
-    
+
+    /* Compute CRC and store as big endian */
+    myCrc = vos_crc32(0, gSampleDATA, 8);
+    gSampleDATA[8]  = (UINT8) myCrc >> 24;
+    gSampleDATA[9]  = (UINT8) myCrc >> 16;
+    gSampleDATA[10] = (UINT8) myCrc >> 8;
+    gSampleDATA[11] = (UINT8) myCrc;
+
     printf("CRC: %08x\n", myCrc);
-    printf("BE in memory (from low to high): %02hhx %02hhx %02hhx %02hhx\n", gSampleDATA[8], gSampleDATA[9], gSampleDATA[10], gSampleDATA[11]);
-    
-	// CRC Check on Null
-    
-	// Verify CRC */
-    myCrc = vos_crc32 (0, gSampleDATA, 12);
+    printf("BE in memory (from low to high): %02hhx %02hhx %02hhx %02hhx\n",
+           gSampleDATA[8],
+           gSampleDATA[9],
+           gSampleDATA[10],
+           gSampleDATA[11]);
+
+    /* CRC Check on Null */
+
+    /* Verify CRC * / */
+    myCrc = vos_crc32(0, gSampleDATA, 12);
     printf("Checked CRC over all: %08x\n", myCrc);
-    
+
     if (myCrc == 0)
     {
         printf(" = Correct behavior!\n");
@@ -82,4 +91,3 @@ int main()
         printf(" = Wrong!!\n");
     }
 }
-
